ctors was made internal, added factory classe, test-classes moved to separate folder
