+++ DIContainer.Test/TestModels/IRepository.cs
@@ -1,0 +1,6 @@
+namespace DIContainer.Test.TestModels
+{
+    public interface IRepository
+    {
+    }
+}
