+++ DIContainer.Test/TestModels/IService.cs
@@ -1,0 +1,7 @@
+namespace DIContainer.Test.TestModels
+{
+    public interface IService
+    {
+        IRepository Repo { get; }
+    }
+}
