+++ DIContainer.Test/TestModels/Repository.cs
@@ -1,0 +1,7 @@
+namespace DIContainer.Test.TestModels
+{
+    public class Repository : IRepository
+    {
+
+    }
+}
