+++ DIContainer/Factories/CollectionFactory.cs
@@ -1,0 +1,10 @@
+﻿namespace DIContainer.Factories
+{
+    public static class CollectionFactory
+    {
+        public static IDependencyCollection CreateCollection()
+        {
+            return new DependencyCollection();;
+        }
+    }
+}
